parkour_app/requirements/base.txt
# SHA1:2e19ca95c9c9f9128934b09ff5c241cd003c0ef7
#
# This file is autogenerated by pip-compile-multi
# To update, run:
#
#    pip-compile-multi
#
alabaster==0.7.12
    # via sphinx
babel==2.10.1
    # via sphinx
bioblend==0.16.0
    # via -r requirements/base.in
boto==2.49.0
    # via bioblend
certifi==2022.5.18.1
    # via requests
charset-normalizer==2.0.12
    # via requests
colorama==0.4.4
    # via sphinx-autobuild
dj-database-url==0.5.0
    # via -r requirements/base.in
django==1.11.29
    # via
    #   -r requirements/base.in
    #   django-authtools
    #   djangorestframework
django-admin-list-filter-dropdown==1.0.3
    # via -r requirements/base.in
django-authtools==1.7.0
    # via -r requirements/base.in
django_extensions==1.9.7
    # via -r requirements/base.in
django-monthfield==0.1.3
    # via -r requirements/base.in
djangorestframework==3.11.2
    # via -r requirements/base.in
docutils==0.17.1
    # via
    #   sphinx
    #   sphinx-rtd-theme
fpdf==1.7.2
    # via -r requirements/base.in
idna==3.3
    # via requests
imagesize==1.3.0
    # via sphinx
importlib-metadata==4.8.3
    # via sphinx
jinja2==3.0.3
    # via sphinx
livereload==2.6.3
    # via sphinx-autobuild
lxml==4.9.0
    # via python-docx
markupsafe==2.0.1
    # via jinja2
numpy==1.19.5
    # via
    #   -r requirements/base.in
    #   pandas
packaging==21.3
    # via sphinx
pandas==1.1.5
    # via -r requirements/base.in
psycopg2-binary==2.8.6
    # via -r requirements/base.in
pygments==2.12.0
    # via sphinx
pyparsing==3.0.9
    # via packaging
python-dateutil==2.8.2
    # via
    #   -r requirements/base.in
    #   pandas
python-docx==0.8.11
    # via -r requirements/base.in
pytz==2022.1
    # via
    #   babel
    #   django
    #   pandas
pyyaml==6.0
    # via bioblend
requests==2.27.1
    # via
    #   bioblend
    #   requests-toolbelt
    #   sphinx
requests-toolbelt==0.9.1
    # via bioblend
six==1.16.0
    # via
    #   django-extensions
    #   livereload
    #   python-dateutil
snowballstemmer==2.2.0
    # via sphinx
sphinx==5.0.1
    # via
    #   -r requirements/base.in
    #   sphinx-autobuild
    #   sphinx-rtd-theme
sphinx-autobuild==2021.3.14
    # via -r requirements/base.in
sphinx-rtd-theme==1.0.0
    # via -r requirements/base.in
sphinxcontrib-applehelp==1.0.2
    # via sphinx
sphinxcontrib-devhelp==1.0.2
    # via sphinx
sphinxcontrib-htmlhelp==2.0.0
    # via sphinx
sphinxcontrib-jsmath==1.0.1
    # via sphinx
sphinxcontrib-qthelp==1.0.3
    # via sphinx
sphinxcontrib-serializinghtml==1.1.5
    # via sphinx
tornado==6.1
    # via livereload
typing==3.7.4.3
    # via django-extensions
typing-extensions==4.1.1
    # via importlib-metadata
urllib3==1.26.9
    # via requests
xlwt==1.3.0
    # via -r requirements/base.in
zipp==3.6.0
    # via importlib-metadata
